requirements.txt
binarytree
Pillow